fix install

+++ Tools/install.ps1
@@ -1,5 +1,4 @@
 param($installPath, $toolsPath, $package, $project)
-
   # Need to load MSBuild assembly if it's not loaded yet.
   Add-Type -AssemblyName 'Microsoft.Build, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a'
   # Grab the loaded MSBuild project for the project
@@ -9,16 +8,20 @@ param($installPath, $toolsPath, $package, $project)
   %{try { $project.ProjectItems.Item('App_Start').ProjectItems.Item('RazorGeneratorMvcStart.cs') } catch { $null }} | ?{$_ -ne $null} | %{ $_.Remove() }
   %{try { $project.ProjectItems.Item('App_Start') } catch { $null }} | ?{$_ -ne $null -and $_.ProjectItems.Count -eq 0} | %{ $_.Remove() }
 
-  %{try { $project.ProjectItems.Item('ResourcePackages').ProjectItems.Item('Bootstrap').ProjectItems.Item('MVC').ProjectItems.Item('Views').ProjectItems.Item('Recaptcha') } catch { $null }} | ?{$_ -ne $null} | %{ $_.Remove() }
+  if (($package.Version -eq "1.4.410.0") -or ((Get-Package "Telerik.Sitefinity.Feather" | where Version -gt "1.4.400.0").Count -eq 0)) {
+    %{try { $project.ProjectItems.Item('ResourcePackages').ProjectItems.Item('Bootstrap').ProjectItems.Item('MVC').ProjectItems.Item('Views').ProjectItems.Item('Recaptcha') } catch { $null }} | ?{$_ -ne $null} | %{ $_.Remove() }
+  }
   
   $project.Save()
 
   $fileInfo = new-object -typename System.IO.FileInfo -ArgumentList $project.FullName
   $projectDirectory = $fileInfo.DirectoryName
   
-  Get-ChildItem "$projectDirectory\App_Start\RazorGeneratorMvcStart.cs" | Remove-Item -Confirm
-  if ((Get-ChildItem "$projectDirectory\App_Start").Length -eq 0) {
-    Remove-Item "$projectDirectory\App_Start"
+  if (Test-Path "$projectDirectory\App_Start\RazorGeneratorMvcStart.cs") {
+	  Get-ChildItem "$projectDirectory\App_Start\RazorGeneratorMvcStart.cs" | Remove-Item
+	  if ((Get-ChildItem "$projectDirectory\App_Start").Length -eq 0) {
+		Remove-Item "$projectDirectory\App_Start"
+	  }
   }
 
   # Make sure all Resource Packages have RazorGenerator directives
@@ -27,8 +30,10 @@ param($installPath, $toolsPath, $package, $project)
     Get-ChildItem "$projectDirectory\ResourcePackages" -Directory -Exclude "Bootstrap" | ?{ $_.GetFiles("razorgenerator.directives").Count -eq 0 } | %{ Copy-Item $generatorDirectivesPath $_.FullName }
   }
 
-  # Prompt to remove Recaptcha template if exists since it isn't distributed with Feather anymore
-  $recaptchaTemplatesPath = "$projectDirectory\ResourcePackages\Bootstrap\MVC\Views\Recaptcha"
-  if (Test-Path $recaptchaTemplatesPath) {
-    Remove-Item $recaptchaTemplatesPath -Recurse -Confirm
+  if (($package.Version -eq "1.4.410.0") -or ((Get-Package "Telerik.Sitefinity.Feather" | where Version -gt "1.4.400.0").Count -eq 0)) {
+      # Prompt to remove Recaptcha template if exists since it isn't distributed with Feather anymore
+      $recaptchaTemplatesPath = "$projectDirectory\ResourcePackages\Bootstrap\MVC\Views\Recaptcha"
+      if (Test-Path $recaptchaTemplatesPath) {
+        Remove-Item $recaptchaTemplatesPath -Recurse
+      }
   }
